+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [Unreleased]
+
+### Added
+
+- Support for `envVars` at the cluster level and task level
+
+### Changed
+
+- Deprecated: `task.environment` has been replaced with `envVars`
+
+
+
 ## [0.4.0] - 2021-05-25
 
 ### Added